+++ src/app/domain/entities/account.py
@@ -1,10 +1,11 @@
 from uuid import UUID
 from dataclasses import dataclass
+from decimal import Decimal
 
 @dataclass
 class Account:
     account_number: str
     user_id: UUID
     name: str
-    balance: float
+    balance: Decimal
     currency: str
